refactor: Service 의존 Repository 변경
