+++ src/main/java/roomescape/member/service/MemberService.java
@@ -1,7 +1,5 @@
 package roomescape.member.service;
 
-import static roomescape.global.exception.ErrorMessage.INTERNAL_SERVER_ERROR;
-
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +9,6 @@ import roomescape.member.domain.Member;
 import roomescape.member.dto.MemberRequest;
 import roomescape.member.dto.MemberResponse;
 import roomescape.member.dto.MemberResponses;
-import roomescape.member.exception.MemberNotFoundException;
 import roomescape.member.repository.MemberRepository;
 
 @Slf4j
@@ -42,21 +39,5 @@ public class MemberService {
                 .toList();
 
         return new MemberResponses(members);
-    }
-
-    public Member getMemberById(final Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> {
-                    log.warn("[회원 조회 실패] 존재하지 않는 회원 ID: {}", memberId);
-                    return new MemberNotFoundException(INTERNAL_SERVER_ERROR.getMessage());
-                });
-    }
-
-    public Member getMemberByEmail(final String email) {
-        return memberRepository.findByEmail(email)
-                .orElseThrow(() -> {
-                    log.warn("[회원 조회 실패] 존재하지 않는 이메일: {}", email);
-                    return new MemberNotFoundException(INTERNAL_SERVER_ERROR.getMessage());
-                });
     }
 }
